backend/requirements.txt
"""TravelSensei Python Flask Backend Requirements
Optimized, deduplicated list (previous file had conflicting duplicate versions).
Pinned for reproducibility; heavy ML libs can be optionally deferred in local dev.
"""

# --- Core Web ---
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Compress==1.14
Werkzeug==2.3.7

# --- Persistence / Config ---
firebase-admin==6.2.0
python-dotenv==1.0.0
pymongo==4.6.0

# --- Data & ML (minimal subset for now; full stack moved to requirements_ai.txt) ---
pandas==2.1.1
scikit-learn==1.3.0
numpy==1.25.2
transformers==4.34.0
torch==2.0.1
sentence-transformers==2.2.2
huggingface-hub==0.17.3

# --- External APIs & Utilities ---
requests==2.31.0
openai==1.3.0
cloudinary==1.36.0
Pillow==10.0.1
python-dateutil==2.8.2

# --- Auth / Validation / Email ---
marshmallow==3.20.1
email-validator==2.0.0
Flask-Mail==0.9.1

# --- Deployment ---
gunicorn==21.2.0

# --- Dev / Testing ---
pytest==7.4.2
pytest-flask==1.2.0